Refactor projects.html layout and update font size and links
This commit refactors the layout of the projects.html page and updates the font size of the date. It also adds line breaks and proper formatting to the links for the works that will be presented at ICML in Vienna. The changes improve the readability and aesthetics of the page.